Add files via upload
This project is User Login/Signup using Embedded JavaScript. In this project, bcrypt is installed for encryption of the data that is to be saved in the database (Mongodb). And then, passport is installed for the authentication of the user. The user's data gets stored in the browser in the form of cookies.